+++ src/components/LeftPanel.tsx
@@ -2,30 +2,19 @@ import { useState, useMemo } from 'react';
 import { usePlaceStore } from '../store/placeStore'; // [추가] 장소 캐시를 사용하기 위해 import
 import { SimpleToggle } from './ui/SimpleToggle';
 import {
-  ListOrdered,
   GripVertical,
   ChevronDown,
-  Lightbulb,
   ChevronUp, // [수정] ChevronsUpDown 아이콘으로 변경
   ChevronsUpDown,
-  MapPin,
   X,
   PlusCircle,
-  Clock,
-  Car,
-  MessageCircle,
   Check,
   RefreshCw,
 } from 'lucide-react';
-import {
-  SortableContext,
-  useSortable,
-  verticalListSortingStrategy,
-} from '@dnd-kit/sortable';
-import { useDroppable } from '@dnd-kit/core';
+import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import type { Poi } from '../hooks/usePoiSocket';
-import type { DayLayer, RouteSegment } from '../types/map';
+import type { DayLayer } from '../types/map';
 import { Button } from './ui/button';
 import React from 'react';
 import { ChatPanel } from './ChatPanel';
@@ -175,331 +164,6 @@ function PoiItem({
     </li>
   );
 }
-
-function MarkerStorage({
-  pois,
-  onPoiClick,
-  onPoiHover,
-  unmarkPoi,
-  removeSchedule,
-  hoveredPoiId,
-}: {
-  pois: Poi[];
-  onPoiClick: (poi: Poi | AiPlace) => void;
-  onPoiHover: (poiId: string | null) => void;
-  unmarkPoi: (poiId: string | number) => void;
-  removeSchedule: (poiId: string, planDayId: string) => void;
-  hoveredPoiId: string | null;
-}) {
-  const { setNodeRef } = useDroppable({ id: 'marker-storage' });
-  const [isCollapsed, setIsCollapsed] = useState(false);
-
-  return (
-    <div ref={setNodeRef} className="p-3 border-b">
-      <div className="flex justify-between items-center mb-2">
-        <div className="flex items-center gap-2">
-          <MapPin className="w-5 h-5 text-gray-600" />
-          <h3 className="text-lg font-bold">장소 보관함</h3>
-        </div>
-        <Button
-          variant="ghost"
-          size="icon"
-          onClick={() => setIsCollapsed(!isCollapsed)}
-        >
-          {isCollapsed ? (
-            <ChevronDown className="h-4 w-4" />
-          ) : (
-            <ChevronUp className="h-4 w-4" />
-          )}
-        </Button>
-      </div>
-      {!isCollapsed && (
-        <SortableContext
-          id="marker-storage-sortable"
-          items={pois.map((p) => p.id)}
-          strategy={verticalListSortingStrategy}
-        >
-          <ul className="space-y-2 min-h-[2rem]">
-            {pois.length > 0 ? (
-              pois.map((poi) => (
-                <PoiItem
-                  key={poi.id}
-                  poi={poi}
-                  onPoiClick={onPoiClick}
-                  onPoiHover={onPoiHover}
-                  unmarkPoi={unmarkPoi}
-                  removeSchedule={removeSchedule}
-                  isHovered={hoveredPoiId === poi.id}
-                />
-              ))
-            ) : (
-              <p className="text-sm text-gray-500 p-2">
-                지도에 장소를 추가하여 보관하세요.
-              </p>
-            )}
-          </ul>
-        </SortableContext>
-      )}
-    </div>
-  );
-}
-
-function DayItineraryItem({
-  layer,
-  itinerary,
-  visibleDayIds,
-  routeSegmentsByDay,
-  onDayVisibilityChange,
-  onOptimizeRoute,
-  onPoiClick,
-  onPoiHover,
-  unmarkPoi,
-  removeSchedule,
-  onToggleCollapse,
-  isCollapsed,
-  hoveredPoiId,
-}: {
-  layer: DayLayer;
-  itinerary: Record<string, Poi[]>;
-  visibleDayIds: Set<string>;
-  routeSegmentsByDay: Record<string, RouteSegment[]>;
-  onDayVisibilityChange: (dayId: string, isVisible: boolean) => void;
-  onOptimizeRoute: (dayId: string) => void;
-  onPoiClick: (poi: Poi | AiPlace) => void;
-  onPoiHover: (poiId: string | null) => void;
-  unmarkPoi: (poiId: string | number) => void;
-  removeSchedule: (poiId: string, planDayId: string) => void;
-  isCollapsed: boolean;
-  onToggleCollapse: () => void;
-  hoveredPoiId: string | null;
-}) {
-  const { setNodeRef } = useDroppable({ id: layer.id });
-
-  const pois = itinerary[layer.id] || [];
-  const isDayVisible = visibleDayIds.has(layer.id);
-  const segmentsForThisDay = routeSegmentsByDay[layer.id] || [];
-  const containerBodyClasses = `transition-opacity duration-300 ${
-    isDayVisible ? 'opacity-100' : 'opacity-40 pointer-events-none'
-  }`;
-
-  return (
-    <div className="border-b pb-2 px-3">
-      <div
-        ref={setNodeRef}
-        className="flex items-center justify-between gap-2 "
-      >
-        <div className="flex items-center gap-2 flex-shrink min-w-0 ">
-          <SimpleToggle
-            checked={isDayVisible}
-            onChange={(checked) => onDayVisibilityChange(layer.id, checked)}
-          />
-          <h3 className="text-base font-bold truncate">{layer.label}</h3>
-        </div>
-        <div className="flex items-center gap-1 flex-shrink-0">
-          {pois.length >= 2 && (
-            <Button
-              variant="outline"
-              size="sm" // size="sm"에 text-sm이 포함되어 있지만 명시적으로 변경
-              className="h-7 text-sm "
-              onClick={() => onOptimizeRoute(layer.id)}
-            >
-              경로 최적화
-            </Button>
-          )}
-          <Button
-            variant="ghost"
-            size="icon"
-            className="flex-shrink-0 "
-            onClick={onToggleCollapse}
-          >
-            {isCollapsed ? (
-              <ChevronDown className="h-4 w-4" />
-            ) : (
-              <ChevronUp className="h-4 w-4" />
-            )}
-          </Button>
-        </div>
-      </div>
-
-      <div className={`mt-2 ${containerBodyClasses}`}>
-        {!isCollapsed && (
-          <SortableContext
-            id={`${layer.id}-sortable`}
-            items={pois.map((p) => p.id)}
-            strategy={verticalListSortingStrategy}
-          >
-            <ul className="space-y-2 min-h-[2rem]">
-              {pois.length > 0 ? (
-                pois.map((poi, index) => (
-                  <React.Fragment key={poi.id}>
-                    <PoiItem
-                      poi={poi}
-                      color={layer.color}
-                      index={index}
-                      onPoiClick={onPoiClick}
-                      onPoiHover={onPoiHover}
-                      unmarkPoi={unmarkPoi}
-                      removeSchedule={removeSchedule}
-                      isHovered={hoveredPoiId === poi.id}
-                    />
-                    {index < pois.length - 1 &&
-                      (() => {
-                        const nextPoi = pois[index + 1];
-                        const segment = segmentsForThisDay.find(
-                          (s) =>
-                            s.fromPoiId === poi.id && s.toPoiId === nextPoi.id
-                        );
-                        if (!segment) return null;
-
-                        const totalMinutes = Math.ceil(segment.duration / 60);
-                        const totalKilometers = (
-                          segment.distance / 1000
-                        ).toFixed(1);
-
-                        return (
-                          <div className="relative flex items-center h-8 pl-8">
-                            <div className="absolute left-4 w-0.5 h-full bg-gray-300" />
-                            <div className="flex items-center text-sm text-gray-600">
-                              <span className="mr-2 flex items-center">
-                                <Clock className="w-3 h-3 mr-1" />
-                                {`${totalMinutes}분`}
-                              </span>
-                              <span className="flex items-center">
-                                <Car className="w-3 h-3 mr-1" />
-                                {`${totalKilometers}km`}
-                              </span>
-                            </div>
-                          </div>
-                        );
-                      })()}
-                  </React.Fragment>
-                ))
-              ) : (
-                <p className="text-sm text-gray-500 p-2">
-                  장소를 보관함에서 드래그하여 추가하세요.
-                </p>
-              )}
-            </ul>
-          </SortableContext>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function ItineraryPanel({
-  itinerary,
-  dayLayers,
-  onPoiClick,
-  onPoiHover,
-  unmarkPoi,
-  removeSchedule,
-  routeSegmentsByDay,
-  onOptimizeRoute,
-  visibleDayIds,
-  onDayVisibilityChange,
-  hoveredPoiId,
-  isRecommendationLoading,
-  onMyItineraryVisibilityChange, // [수정] '내 일정' 전체 토글 핸들러 prop
-}: {
-  workspaceId: string;
-  isRecommendationLoading: boolean;
-  itinerary: Record<string, Poi[]>;
-  dayLayers: DayLayer[];
-  onPoiClick: (poi: Poi | AiPlace) => void;
-  onPoiHover: (poiId: string | null) => void;
-  unmarkPoi: (poiId: string | number) => void;
-  removeSchedule: (poiId: string, planDayId: string) => void;
-  routeSegmentsByDay: Record<string, RouteSegment[]>;
-  onOptimizeRoute: (dayId: string) => void;
-  visibleDayIds: Set<string>;
-  onDayVisibilityChange: (dayId: string, isVisible: boolean) => void;
-  hoveredPoiId: string | null;
-  onMyItineraryVisibilityChange: () => void; // [수정] '내 일정' 전체 토글 핸들러 prop
-}) {
-  // [신규] '내 일정'의 접기/펼치기 상태 관리
-  const [collapsedDayIds, setCollapsedDayIds] = useState<Set<string>>(
-    new Set()
-  );
-
-  const handleToggleDayCollapse = (dayId: string) => {
-    setCollapsedDayIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(dayId)) {
-        newSet.delete(dayId);
-      } else {
-        newSet.add(dayId);
-      }
-      return newSet;
-    });
-  };
-
-  const handleToggleAllCollapse = () => {
-    // 모든 날짜가 접혀있으면 모두 펴고, 그렇지 않으면 모두 접습니다.
-    if (collapsedDayIds.size === dayLayers.length) {
-      setCollapsedDayIds(new Set());
-    } else {
-      setCollapsedDayIds(new Set(dayLayers.map((l) => l.id)));
-    }
-  };
-
-  return (
-    <div className="p-3 space-y-3 flex flex-col">
-      {isRecommendationLoading ? (
-        <div className="flex justify-center items-center h-full text-base text-gray-500">
-          AI 추천 일정을 불러오는 중...
-        </div>
-      ) : (
-        <>
-          {/* [신규] 전체 경로 토글 스위치 */}
-          <div className="flex items-center justify-between p-2 border-b">
-            <div className="flex items-center gap-2">
-              <h3 className="text-base font-bold">전체 경로</h3>
-              <SimpleToggle
-                // '내 일정'의 모든 경로가 켜져 있을 때만 ON
-                checked={dayLayers.every((layer) =>
-                  visibleDayIds.has(layer.id)
-                )}
-                onChange={onMyItineraryVisibilityChange}
-              />
-            </div>
-            {/* [신규] 모두 접기/펴기 버튼 */}
-            <Button
-              variant="link"
-              size="sm"
-              className="text-base text-gray-500"
-              onClick={handleToggleAllCollapse}
-            >
-              <ChevronsUpDown className="w-3.5 h-3.5 mr-1" />
-              {collapsedDayIds.size === dayLayers.length
-                ? '일정 모두 펴기'
-                : '일정 모두 접기'}
-            </Button>
-          </div>
-          {dayLayers.map((layer) => (
-            <DayItineraryItem
-              isCollapsed={collapsedDayIds.has(layer.id)} // [수정] isCollapsed prop 전달
-              key={layer.id}
-              layer={layer}
-              itinerary={itinerary}
-              visibleDayIds={visibleDayIds}
-              routeSegmentsByDay={routeSegmentsByDay}
-              onDayVisibilityChange={onDayVisibilityChange}
-              onOptimizeRoute={onOptimizeRoute}
-              onPoiClick={onPoiClick}
-              onPoiHover={onPoiHover}
-              unmarkPoi={unmarkPoi}
-              removeSchedule={removeSchedule}
-              onToggleCollapse={() => handleToggleDayCollapse(layer.id)}
-              hoveredPoiId={hoveredPoiId}
-            />
-          ))}
-        </>
-      )}
-    </div>
-  );
-}
-
 interface LeftPanelProps {
   isOpen: boolean;
   isRecommendationLoading: boolean;
@@ -521,228 +185,6 @@ interface LeftPanelProps {
   onCardClick: (poi: any) => void;
   setChatAiPlaces: (places: AiPlace[]) => void;
   chatAiPlaces: AiPlace[];
-}
-
-const formatDuration = (seconds: number) => {
-  const minutes = Math.ceil(seconds / 60);
-  if (minutes < 60) return `${minutes}분`;
-  const hours = Math.floor(minutes / 60);
-  const remainingMinutes = minutes % 60;
-  return `${hours}시간 ${remainingMinutes}분`;
-};
-
-const formatDistance = (meters: number) => {
-  return `${(meters / 1000).toFixed(1)}km`;
-};
-
-const formatDurationChange = (seconds: number) => {
-  const isNegative = seconds < 0;
-  const prefix = isNegative ? '' : '+';
-  const absSeconds = Math.abs(seconds);
-  const minutes = Math.floor(absSeconds / 60);
-
-  if (minutes === 0 && absSeconds > 0) {
-    return seconds > 0 ? '+1분 미만' : '-1분 미만';
-  }
-
-  let formatted;
-  if (minutes < 60) {
-    formatted = `${minutes}분`;
-  } else {
-    const hours = Math.floor(minutes / 60);
-    const remainingMinutes = minutes % 60;
-    formatted = `${hours}시간 ${remainingMinutes}분`;
-  }
-  if (seconds === 0) return '변동 없음';
-  return `${prefix}${formatted}`;
-};
-
-const formatDistanceChange = (meters: number) => {
-  const isNegative = meters < 0;
-  const prefix = isNegative ? '' : '+';
-  const absMeters = Math.abs(meters);
-  const formatted = `${(absMeters / 1000).toFixed(1)}km`;
-  if (meters === 0) return '변동 없음';
-  return `${prefix}${formatted}`;
-};
-
-interface OptimizationModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  originalData: { pois: Poi[]; segments: RouteSegment[] } | null;
-  optimizedData: { pois: Poi[]; segments: RouteSegment[] } | null;
-  dayLayer: DayLayer | null;
-}
-
-function OptimizationModal({
-  isOpen,
-  onClose,
-  originalData,
-  optimizedData,
-  dayLayer,
-}: OptimizationModalProps) {
-  if (!isOpen || !originalData || !dayLayer) return null;
-
-  const calculateTotals = (segments: RouteSegment[]) => {
-    const totalDistance = segments.reduce((sum, s) => sum + s.distance, 0);
-    const totalDuration = segments.reduce((sum, s) => sum + s.duration, 0);
-    return { totalDistance, totalDuration };
-  };
-
-  const originalTotals = calculateTotals(originalData.segments);
-  const optimizedTotals = optimizedData
-    ? calculateTotals(optimizedData.segments)
-    : null;
-
-  const renderRouteList = (
-    pois: Poi[],
-    segments: RouteSegment[],
-    color: string
-  ) => (
-    <ul className="space-y-1">
-      {pois.map((poi, index) => (
-        <React.Fragment key={poi.id}>
-          <li className="flex items-center text-base">
-            <span
-              className="flex-shrink-0 flex items-center justify-center w-5 h-5 rounded-full text-white text-sm mr-3"
-              style={{ backgroundColor: color }}
-            >
-              {index + 1}
-            </span>
-            <span className="truncate">{poi.placeName}</span>
-          </li>
-          {index < pois.length - 1 &&
-            (() => {
-              const nextPoi = pois[index + 1];
-              if (!nextPoi) return null;
-              const segment = segments.find(
-                (s) => s.fromPoiId === poi.id && s.toPoiId === nextPoi.id
-              );
-              if (!segment) return null;
-              return (
-                <div className="relative flex h-8 items-center pl-8">
-                  <div className="absolute left-4 h-full w-0.5 bg-gray-300" />
-                  <div className="flex items-center text-sm text-gray-600">
-                    <span className="mr-2 flex items-center">
-                      <Clock className="h-3 w-3 mr-1" />
-                      {formatDuration(segment.duration)}
-                    </span>
-                    <span className="flex items-center">
-                      <Car className="h-3 w-3 mr-1" />
-                      {formatDistance(segment.distance)}
-                    </span>
-                  </div>
-                </div>
-              );
-            })()}
-        </React.Fragment>
-      ))}
-    </ul>
-  );
-
-  return (
-    <div className="fixed inset-0 bg-black/30 flex justify-center items-center z-50">
-      <div className="bg-white rounded-lg shadow-xl w-full max-w-3xl">
-        <div className="p-4 border-b flex justify-between items-center">
-          <h2 className="text-xl font-bold">경로 최적화 결과</h2>
-          <Button variant="ghost" size="icon" onClick={onClose}>
-            <X className="w-4 h-4" />
-          </Button>
-        </div>
-        <div className="grid grid-cols-2 divide-x">
-          <div className="p-4">
-            <h3 className="font-semibold mb-3 text-center">기존 경로</h3>
-            <div className="text-base mb-4 p-2 bg-gray-50 rounded-md">
-              <p>
-                <strong>총 거리:</strong>{' '}
-                {formatDistance(originalTotals.totalDistance)}
-              </p>
-              <p>
-                <strong>총 소요 시간:</strong>{' '}
-                {formatDuration(originalTotals.totalDuration)}
-              </p>
-            </div>
-            <div className="max-h-80 overflow-y-auto pr-2">
-              {renderRouteList(
-                originalData.pois,
-                originalData.segments,
-                dayLayer.color
-              )}
-            </div>
-          </div>
-
-          <div className="p-4">
-            <h3 className="font-semibold mb-3 text-center">최적 경로</h3>
-            {optimizedData && optimizedTotals ? (
-              <>
-                <div className="text-base mb-4 p-2 bg-blue-50 rounded-md">
-                  <p>
-                    <strong>총 거리:</strong>{' '}
-                    {formatDistance(optimizedTotals.totalDistance)}
-                    <span
-                      className={`ml-2 text-sm font-semibold ${
-                        optimizedTotals.totalDistance <
-                        originalTotals.totalDistance
-                          ? 'text-blue-600'
-                          : optimizedTotals.totalDistance >
-                              originalTotals.totalDistance
-                            ? 'text-red-600'
-                            : 'text-gray-500'
-                      }`}
-                    >
-                      (
-                      {formatDistanceChange(
-                        optimizedTotals.totalDistance -
-                          originalTotals.totalDistance
-                      )}
-                      )
-                    </span>
-                  </p>
-                  <p>
-                    <strong>총 소요 시간:</strong>{' '}
-                    {formatDuration(optimizedTotals.totalDuration)}
-                    <span
-                      className={`ml-2 text-sm font-semibold ${
-                        optimizedTotals.totalDuration <
-                        originalTotals.totalDuration
-                          ? 'text-blue-600'
-                          : optimizedTotals.totalDuration >
-                              originalTotals.totalDuration
-                            ? 'text-red-600'
-                            : 'text-gray-500'
-                      }`}
-                    >
-                      (
-                      {formatDurationChange(
-                        optimizedTotals.totalDuration -
-                          originalTotals.totalDuration
-                      )}
-                      )
-                    </span>
-                  </p>
-                </div>
-                <div className="max-h-80 overflow-y-auto pr-2">
-                  {renderRouteList(
-                    optimizedData.pois,
-                    optimizedData.segments,
-                    dayLayer.color
-                  )}
-                </div>
-              </>
-            ) : (
-              <div className="flex h-full flex-col items-center justify-center">
-                <div className="h-12 w-12 animate-spin rounded-full border-b-2 border-gray-900"></div>
-                <p className="mt-4 text-gray-600">최적화 중입니다...</p>
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="p-4 border-t text-right">
-          <Button onClick={onClose}>닫기</Button>
-        </div>
-      </div>
-    </div>
-  );
 }
 
 function RecommendedDayItem({
@@ -880,7 +322,6 @@ function ChatSidebar({
 
 export function LeftPanel({
   isOpen,
-  isRecommendationLoading,
   workspaceId,
   recommendedItinerary,
   dayLayers,
@@ -900,13 +341,13 @@ export function LeftPanel({
   setChatAiPlaces,
   chatAiPlaces,
 }: LeftPanelProps) {
-  const [activeTab, setActiveTab] = useState('chat');
+  const [activeTab, _setActiveTab] = useState('chat');
 
   const [recCollapsedDayIds, setRecCollapsedDayIds] = useState<Set<string>>(
     new Set()
   );
 
-  const [isAiRecOpen, setIsAiRecOpen] = useState(false);
+  const [isAiRecOpen, _setIsAiRecOpen] = useState(false);
 
   const handleRecToggleDayCollapse = (dayId: string) => {
     setRecCollapsedDayIds((prev) => {
